+++ packages/root/types/navigationConfirmation.d.ts
@@ -1,2 +1,2 @@
 import { ConfirmationMethods } from "./types/navigationConfirmation";
-export default function createNavigationConfirmation(): ConfirmationMethods;
+export default function createNavigationConfirmation<Q>(): ConfirmationMethods<Q>;