+++ packages/root/types/types/index.d.ts
@@ -1,4 +1,4 @@
-export { LocationDetails, PartialLocation, HickoryLocation, AnyLocation } from "./location";
+export { LocationDetails, PartialLocation, HickoryLocation, AnyLocation, KeylessLocation } from "./location";
 export { History, ToArgument, Action, NavType, PendingNavigation, ResponseHandler, Options, CommonHistory } from "./hickory";
 export { KeyMethods, KeyFns } from "./keygen";
 export { QueryFunctions, LocationFactoryOptions, LocationMethods } from "./locationFactory";