+++ packages/root/types/types/navigationConfirmation.d.ts
@@ -1,13 +1,13 @@
 import { HickoryLocation } from "./location";
 import { Action } from "./hickory";
-export interface NavigationInfo {
-    to: HickoryLocation;
-    from: HickoryLocation;
+export interface NavigationInfo<Q> {
+    to: HickoryLocation<Q>;
+    from: HickoryLocation<Q>;
     action: Action;
 }
-export declare type ConfirmationFunction = (info: NavigationInfo, confirm: () => void, prevent?: () => void) => void;
-export interface ConfirmationMethods {
-    confirmNavigation(info: NavigationInfo, confirm: () => void, prevent?: () => void): void;
-    confirmWith(fn?: ConfirmationFunction): void;
+export declare type ConfirmationFunction<Q> = (info: NavigationInfo<Q>, confirm: () => void, prevent?: () => void) => void;
+export interface ConfirmationMethods<Q> {
+    confirmNavigation(info: NavigationInfo<Q>, confirm: () => void, prevent?: () => void): void;
+    confirmWith(fn?: ConfirmationFunction<Q>): void;
     removeConfirmation(): void;
 }